requirements.txt
Flask==2.3.3
Flask-Dance==7.0.0
Flask-SQLAlchemy==3.0.5
# psycopg2-binary==2.9.7  # Uncomment for PostgreSQL production
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
requests==2.31.0 